refactor: 파일명 closeResultModalEvent에서 불필요한 Event suffix 제외
