+++ src/view/web/modules/closeResultModal.js
